Refactored converters and added `ConverterHelper` to simplify usage
