+++ src/Converters/JsonConverter.php
@@ -11,8 +11,10 @@ use Illuminate\Container\Attributes\Config;
 
 use function is_array;
 use function json_encode;
+use function mb_substr;
+use function sprintf;
 
-class ConvertToJsonLines extends Converter
+class JsonConverter extends Converter
 {
     public function __construct(
         #[Config('feeds.converters.json.options')]
@@ -22,37 +24,43 @@ class ConvertToJsonLines extends Converter
         TransformerService $transformer
     ) {
         parent::__construct($pretty, $transformer);
-
-        $this->options &= ~JSON_PRETTY_PRINT;
     }
 
     public function header(Feed $feed): string
     {
-        return '';
+        return $feed->root()->name ? "{\n" : "[\n";
     }
 
     public function footer(Feed $feed): string
     {
-        return '';
+        return $feed->root()->name ? "\n]\n}\n" : "\n]\n";
     }
 
     public function root(Feed $feed): string
     {
-        return '';
+        return sprintf("\"%s\": [\n", $feed->root()->name);
     }
 
     public function item(FeedItem $item, bool $isLast): string
     {
         $data = $this->performItem($item->toArray());
 
-        return $this->toJson($data);
+        $suffix = $isLast ? '' : ',';
+
+        return $this->toJson($data) . $suffix;
     }
 
     public function info(array $info, bool $afterRoot): string
     {
         $data = $this->performItem($info);
 
-        return $this->toJson($data);
+        $json = $this->toJson($data);
+
+        if (! $afterRoot) {
+            $json = mb_substr($json, 1, -1);
+        }
+
+        return $json . ',';
     }
 
     protected function performItem(array $data): array
@@ -72,6 +80,15 @@ class ConvertToJsonLines extends Converter
 
     protected function toJson(array $data): string
     {
-        return json_encode($data, $this->options);
+        return json_encode($data, $this->jsonOptions());
+    }
+
+    protected function jsonOptions(): int
+    {
+        if ($this->pretty) {
+            return JSON_PRETTY_PRINT | $this->options;
+        }
+
+        return $this->options;
     }
 }
