+++ tests/Unit/Architecture/ConverterTest.php
@@ -1,0 +1,7 @@
+<?php
+
+declare(strict_types=1);
+
+arch()
+    ->expect('DragonCode\LaravelFeed\Converters')
+    ->toHaveSuffix('Converter');
